(BSR)[API] refactor: move tests from models/offerer_test.py to core/offerers/test_models.py
